added ControllerAdvice and Bean Validation on DTos

+++ bootcampspringboot/src/main/java/com/max/bootcampspringboot/api/EmployeeController.java
@@ -3,12 +3,16 @@ package com.max.bootcampspringboot.api;
 import com.max.bootcampspringboot.api.mapper.ApiEmployeeMapper;
 import com.max.bootcampspringboot.api.model.ApiEmployee;
 import com.max.bootcampspringboot.service.EmployeeService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/employees")
+@Validated
 public class EmployeeController {
     private final EmployeeService employeeService;
     public EmployeeController(EmployeeService employeeService) {
@@ -26,8 +30,8 @@ public class EmployeeController {
     }
 
     @PostMapping
-    public ApiEmployee addEmployee(@RequestBody ApiEmployee employee){
-        return ApiEmployeeMapper.toApiEmployee(employeeService.addEmployee(ApiEmployeeMapper.toServiceEmployee(employee)));
+    public ResponseEntity<ApiEmployee> addEmployee(@RequestBody ApiEmployee employee){
+        return ResponseEntity.status(HttpStatus.CREATED).body(ApiEmployeeMapper.toApiEmployee(employeeService.addEmployee(ApiEmployeeMapper.toServiceEmployee(employee))));
     }
 
     @PutMapping
@@ -36,8 +40,13 @@ public class EmployeeController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteEmployee(@PathVariable int id){
+    public ResponseEntity<String> deleteEmployee(@PathVariable int id){
+        ApiEmployee temp =ApiEmployeeMapper.toApiEmployee(employeeService.getEmployee(id));
+
+        if (temp == null) throw new RuntimeException("Employee id not found");
         employeeService.deleteEmployee(id);
+
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("");
     }
 
     @GetMapping("/firstnames")
