+++ bootcampspringboot/src/main/java/com/max/bootcampspringboot/api/SkillController.java
@@ -3,6 +3,9 @@ package com.max.bootcampspringboot.api;
 import com.max.bootcampspringboot.api.mapper.ApiSkillMapper;
 import com.max.bootcampspringboot.api.model.ApiSkill;
 import com.max.bootcampspringboot.service.SkillService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -10,6 +13,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/skills")
+@Validated
 public class SkillController {
     private final SkillService skillService;
     public SkillController(SkillService skillService) {
@@ -27,8 +31,8 @@ public class SkillController {
     }
 
     @PostMapping
-    public ApiSkill addSkill(@RequestBody ApiSkill skill){
-        return ApiSkillMapper.toApiSkill(skillService.addSkill(ApiSkillMapper.toServiceSkill(skill)));
+    public ResponseEntity<ApiSkill> addSkill( @RequestBody ApiSkill skill){
+        return ResponseEntity.status(HttpStatus.CREATED).body(ApiSkillMapper.toApiSkill(skillService.addSkill(ApiSkillMapper.toServiceSkill(skill))));
     }
 
     @PutMapping
@@ -37,8 +41,14 @@ public class SkillController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteSkill(@PathVariable int id){
+    public ResponseEntity<String> deleteSkill(@PathVariable int id){
+        ApiSkill temp = ApiSkillMapper.toApiSkill(skillService.getSkill(id));
+
+        if (temp == null) throw new RuntimeException("Skill id not found");
         skillService.deleteSkill(id);
+
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("");
+
     }
 
     @GetMapping("/rarest-skill")
